Refactor CSV row-specific methods to their own class
+ Also rename X2ExportCsvTransformer to plain old CsvTransformer
+ It can handle CSVs from other sources besides X2, shorter is better
